Check for Git, Curl and Vim on install.sh
If the system does not have Git or Curl, the script will refuse to
do its job, since Git is required to clone the repository and Curl
is required to download vim-plug (although wget could be considered
as an alternative).

If Vim is not present on the system, the script will instead provide
a quick help guide on what does he have to run on the command line
once Vim is installed.

+++ install.sh
@@ -1,6 +1,18 @@
 #!/bin/sh
 # vimrc install script
 
+# Check that Git is actually present
+if ! git --version >/dev/null 2>&1; then
+    echo "You need to install Git to clone the vimrc repository."
+    exit 1
+fi
+
+# Check that Curl is present too
+if ! curl --version >/dev/null 2>&1; then
+    echo "You need to install Curl to download vim-plug."
+    exit 1
+fi
+
 # Backup old setup (-f for overwriting old backups).
 [ -e ~/.vimrc ] && (rm -rf ~/.vimrc.bak; mv ~/.vimrc ~/.vimrc.bak)
 [ -e ~/.gvimrc ] && (rm -rf ~/.gvimrc.bak; mv ~/.gvimrc ~/.gvimrc.bak)
@@ -17,5 +29,11 @@ ln -s ~/.vim/vimrc ~/.vimrc
 ln -s ~/.vim/gvimrc ~/.gvimrc
 
 # Execute install.
-vim +PlugInstall +qall
+if vim --version >/dev/null 2>&1; then
+    vim +PlugInstall +qall
+else
+    echo "You don't have Vim installed at this moment."
+    echo "Please, install it, then run:"
+    echo "  vim +PlugInstall +qall"
+fi
